@@ -2,10 +2,12 @@
   "name": "angular-rest",
   "version": "0.0.1",
   "peerDependencies": {
-    "@angular/common": "^6.0.0-rc.0 || ^6.0.0",
-    "@angular/core": "^6.0.0-rc.0 || ^6.0.0"
+    "@angular/common": "^6.1.0",
+    "@angular/core": "^6.1.0"
   },
   "dependencies": {
+    "@angular/common": "^6.1.0",
+    "@angular/core": "^6.1.0",
     "@angular-redux/form": "^9.0.1",
     "@angular-redux/router": "^9.0.0",
     "@angular-redux/store": "^9.0.0",
@@ -16,8 +18,8 @@
     "redux-observable": "^1.0.0"
   },
   "devDependencies": {
-    "@angular/common": "^6.0.0-rc.0 || ^6.0.0",
-    "@angular/core": "^6.0.0-rc.0 || ^6.0.0",
+    "@angular/common": "^6.1.0",
+    "@angular/core": "^6.1.0",
     "@commitlint/cli": "^7.0.0",
     "@commitlint/config-conventional": "^7.0.1",
     "tslint": "^5.11.0"
