@@ -1,9 +1,9 @@
-import { NgModule, ModuleWithProviders, Injector } from '@angular/core';
-import { AngularRestComponent } from './angular-rest.component';
-import { NgRedux, NgReduxModule } from '@angular-redux/store';
-import { IAppState, RootReducer } from './stores/root.store';
-import { NgReduxRouterModule } from '@angular-redux/router';
-import { HttpClientModule } from '@angular/common/http';
+import {Injector, NgModule} from '@angular/core';
+import {AngularRestComponent} from './angular-rest.component';
+import {NgRedux, NgReduxModule} from '@angular-redux/store';
+import {IAppState, RootReducer} from './stores/root.store';
+import {NgReduxRouterModule} from '@angular-redux/router';
+import {HttpClientModule} from '@angular/common/http';
 
 @NgModule({
   imports: [
