@@ -1,8 +1,9 @@
 import {entitiesList} from '../domain/entities.list';
-import {EntityDescriptor} from "../domain/descriptors";
+import {EntityDescriptor} from '../domain/descriptors';
 
-export function Entity(entityDescriptor: EntityDescriptor): ClassDecorator {
-    return function (constructor: any) {
+export function Entity<T extends EntityDescriptor>(entityDescriptor: T): ClassDecorator {
+  return function (constructor: any) {
+    console.log(constructor);
         entitiesList.push(entityDescriptor);
     };
 }
